[ASTextNode] call initialize after possibly dynamically changing the superclass (#1970)
* [ASTextNode] call initialize after possibly dynamically changing the superclass

In order to experiment with ASTN2 we do some crazy obj-c voodoo to change the superclass of an object from `ASTextNode` to `ASTextNode2`. However, before we do that we set up the object’s `_flags` which, among other things, caches whether or not a class responds to certain selectors.

ASTN and ASTN2 draw themselves differently, with ASTN using `displayWithParameters:isCancelled:` and ASTN2 using `drawRect:withParameters:isCancelled:isRasterizing:`. If we are experimenting with ASTN2, if we set up `_flags` before teh swap our eventual ASTN2 class will think it responds to `displayWithParameters:isCancelled:`. If we wait until after the possible swap to setup `_flags` everything works as expected!.

* i want to run the non carthage tests…

* [ASTN2] Fix the podspec (#1971)

* [ASTN2] Fix the podspec

I was too aggressive in including everything from `TextExperiment` as a public header file. `ASTextUtilities.h` includes a file (`ASInternalHelpers.h`) that isn’t exported as a public header and causes build issues. Hopefully this is the proper includes!

* don’t test carthage

* carthage is back

* builds passed, putting carthage back in