Don't mirror isAccessibilityElement on _ASTableViewCell
This is a partial revert of #1941

The change in #1941 has caused some issues for us and others (see #1997).
If I set `myNode.isAccessibilityElement = true,` then what ends up
happening is that the table cell wrapper view (_ASTableViewCell) also
has `isAccessibilityElement = true`. But then when iOS goes through the
hierarchy it identifies the table cell wrapper view as the target and
attempts to use that accessibilityLabel, which doesn't exist
(because it's actually on the nested node view).
So our node's label is never used.

I'm not sure why setting `isAccessibilityElement` in `_ASTableViewCell`
was necessary in #1941 in order to make a node non-accessible,
since the default value is already false.